+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+## Version 0.3 (Sept 8, 2019 JST)
+
+- support multiple break operands (Destructive change)
+  - remove function: `setBreakOperand`
+  - add functions: `addBreakOperand` , `removeBreakOperand` , `removeAllBreakOperands`
+
 ## Version 0.2 (Sept 8, 2019 JST)
 
 - support multiple break points (Destructive change)